pyproject.toml
[project]
name = "ai-trading-agent"
version = "0.1.0"
description = "AI trading agent for Capital.com demo-first execution, with modular decision engine, forecasting, logging, and future multi-broker support."
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary",
    "sqlalchemy",
    "python-dotenv",
    "httpx",
    "pandas",
    "numpy",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-asyncio",
    "ruff",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
